\name{erfe-package}
\alias{erfe-package}
\alias{erfe}
\docType{package}
\title{
\packageTitle{erfe}
}
\description{
\packageDescription{erfe}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{erfe}
\packageIndices{erfe}
This package provide functions to estimate the erfe estimator which is a generalisation of the within estimator using expectile.
}
\author{ Amadou Barry}
\references{
See the documentation for Armadillo, and RcppArmadillo, for more details.
}

\keyword{ package }

\examples{
  x <- sqrt(1:4)
  rcpparma_innerproduct(x)
  rcpparma_outerproduct(x)
}
